Refactor tests into proper source files
